Extract persistence to interface and add a wordpress options implementation
