+++ modules/ppcp-wc-gateway/services.php
@@ -80,8 +80,8 @@ use WooCommerce\PayPalCommerce\WcGateway\Notice\UnsupportedCurrencyAdminNotice;
 use WooCommerce\PayPalCommerce\WcGateway\Processor\AuthorizedPaymentsProcessor;
 use WooCommerce\PayPalCommerce\WcGateway\Processor\OrderProcessor;
 use WooCommerce\PayPalCommerce\WcGateway\Processor\RefundProcessor;
-use WooCommerce\PayPalCommerce\WcGateway\Service\FailedOrderTracker;
-use WooCommerce\PayPalCommerce\WcGateway\Service\WordPressOptionsFailedOrderPersistence;
+use WooCommerce\PayPalCommerce\WcGateway\Service\FailedOrders\FailedOrderTracker;
+use WooCommerce\PayPalCommerce\WcGateway\Service\FailedOrders\WordPressOptionsFailedOrderPersistence;
 use WooCommerce\PayPalCommerce\WcGateway\Settings\HeaderRenderer;
 use WooCommerce\PayPalCommerce\WcGateway\Settings\SectionsRenderer;
 use WooCommerce\PayPalCommerce\WcGateway\Settings\Settings;
